Create README.md
The Double-Double Laminate Finder is a computer tool developed based on Classical Lamination Theory (CLT) to find Double-Double (DD) composite layups that are equivalent in in-plane stiffness [A] or flexural stiffness [D] to conventional quadriaxial (QUAD) layups. This tool can also be used to evaluate material homogenization conditions, determine the optimal DD stacking sequence for homogenization, and indicate the QUAD and DD layups in the in-plane and flexural design range plots.

The code is written in Python and runs on computers using the Windows operating system.